feat(web): add the linking of the discord integration
