+++ web/src/app/features/home/home.page.html
@@ -1,4 +1,3 @@
 <div class="flex flex-col">
   <h1 class="tr-title-1">Home</h1>
-  <button trBtn (click)="openDiscordOAuthPage()">Connect to Discord</button>
 </div>
