+++ web/src/app/features/home/home.page.ts
@@ -1,7 +1,4 @@
-import { ChangeDetectionStrategy, Component, inject, signal } from '@angular/core';
-import { TrButtonDirective } from '@triggo-ui/button';
-import { finalize } from 'rxjs/operators';
-import { DiscordRepository } from '@repositories/integrations';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 @Component({
   selector: 'tr-home',
@@ -9,33 +6,6 @@ import { DiscordRepository } from '@repositories/integrations';
   standalone: true,
   styles: [],
   changeDetection: ChangeDetectionStrategy.OnPush,
-  imports: [TrButtonDirective],
+  imports: [],
 })
-export class HomePageComponent {
-  readonly #discordRepository = inject(DiscordRepository);
-
-  loading = signal<boolean>(false);
-  error = signal<boolean>(false);
-
-  openDiscordOAuthPage() {
-    const url = `${window.location.origin}/oauth2/discord`;
-    this.#discordRepository
-      .getUri()
-      .pipe(finalize(() => this.loading.set(false)))
-      .subscribe({
-        next: uri => {
-          if (!uri) {
-            this.error.set(true);
-            return;
-          }
-          const newWindow = window.open(`${uri}`, '_blank');
-          if (newWindow) {
-            newWindow.opener = window;
-          }
-        },
-        error: () => {
-          this.error.set(true);
-        },
-      });
-  }
-}
+export class HomePageComponent {}
