+++ web/src/app/features/integrations/components/integration-available-card/integration-available-card.component.ts
@@ -6,6 +6,7 @@ export interface IntegrationAvailableCardProps {
   name: string;
   description: string;
   features: string[];
+  linkFn: () => void;
 }
 
 @Component({
